+++ flowman-core/src/main/scala/com/dimajix/flowman/execution/Session.scala
@@ -19,6 +19,7 @@ package com.dimajix.flowman.execution
 import org.apache.hadoop.conf.Configuration
 import org.apache.spark.SparkConf
 import org.apache.spark.sql.SparkSession
+import org.apache.spark.sql.SparkShim
 import org.apache.spark.sql.internal.SQLConf
 import org.slf4j.LoggerFactory
 
@@ -267,7 +268,7 @@ class Session private[execution](
                 logger.info("Creating Spark session using provided builder")
                 // Set all session properties that can be changed in an existing session
                 sparkConf.getAll.foreach { case (key, value) =>
-                    if (!SQLConf.staticConfKeys.contains(key)) {
+                    if (!SparkShim.isStaticConf(key)) {
                         injectedSession.conf.set(key, value)
                     }
                 }