+++ flowman-core/src/main/scala/com/dimajix/flowman/spec/model/HiveTableRelation.scala
@@ -22,12 +22,12 @@ import com.fasterxml.jackson.annotation.JsonProperty
 import org.apache.hadoop.fs.Path
 import org.apache.spark.sql.DataFrame
 import org.apache.spark.sql.catalyst.TableIdentifier
-import org.apache.spark.sql.catalyst.analysis.UnresolvedRelation
 import org.apache.spark.sql.catalyst.catalog.CatalogStorageFormat
 import org.apache.spark.sql.catalyst.catalog.CatalogTable
 import org.apache.spark.sql.catalyst.catalog.CatalogTableType
-import org.apache.spark.sql.catalyst.plans.logical.InsertIntoTable
+import org.apache.spark.sql.catalyst.expressions.Attribute
 import org.apache.spark.sql.execution.SQLExecution
+import org.apache.spark.sql.hive.execution.InsertIntoHiveTable
 import org.apache.spark.sql.internal.HiveSerDe
 import org.apache.spark.sql.types.StructType
 import org.slf4j.LoggerFactory
@@ -146,21 +146,28 @@ class HiveTableRelation(
         // Apply output schema before writing to Hive
         val outputDf = applyOutputSchema(executor, df)
 
+        // Helper method for Spark < 2.4
+        implicit def toAttributeNames(atts:Seq[Attribute]) : Seq[String] = atts.map(_.name)
+
         if (partitionSpec.nonEmpty) {
             val spark = executor.spark
+            val catalog = executor.catalog
+            val hiveTable = catalog.getTable(TableIdentifier(table, database))
+            val query = outputDf.queryExecution.logical
 
             val overwrite = mode.toLowerCase(Locale.ROOT) == "overwrite"
-            val cmd = InsertIntoTable(
-                table = UnresolvedRelation(TableIdentifier(table, database)),
+            val cmd = InsertIntoHiveTable(
+                table = hiveTable,
                 partition = partitionSpec.toMap.mapValues(v => Some(v.toString)),
-                query = outputDf.queryExecution.logical,
+                query = query,
                 overwrite = overwrite,
-                ifPartitionNotExists = false)
+                ifPartitionNotExists = false,
+                query.output
+            )
             val qe = spark.sessionState.executePlan(cmd)
             SQLExecution.withNewExecutionId(spark, qe)(qe.toRdd)
 
             // Finally add Hive partition
-            val catalog = executor.catalog
             val location = catalog.getPartitionLocation(tableIdentifier, partitionSpec)
             catalog.addOrReplacePartition(tableIdentifier, partitionSpec, location)
         }
@@ -223,7 +230,7 @@ class HiveTableRelation(
         require(executor != null)
         require(partitions != null)
 
-        logger.info(s"Cleaning Hive table '$tableIdentifier' for relatio '$identifier'")
+        logger.info(s"Cleaning Hive table '$tableIdentifier' for relation '$identifier'")
 
         val catalog = executor.catalog
         // When no partitions are specified, this implies that the whole table is to be truncated
@@ -247,7 +254,7 @@ class HiveTableRelation(
         require(executor != null)
 
         if (!ifNotExists || !exists(executor)) {
-            logger.info(s"Creating Hive table '$tableIdentifier' for relation '$identifier'")
+            logger.info(s"Creating Hive table $tableIdentifier for relation '$identifier'")
 
             // Create and save Avro schema
             import HiveTableRelation._
@@ -325,7 +332,7 @@ class HiveTableRelation(
 
         val catalog = executor.catalog
         if (!ifExists || catalog.tableExists(tableIdentifier)) {
-            logger.info(s"Destroying Hive table '$tableIdentifier' of relation '$identifier'")
+            logger.info(s"Destroying Hive table $tableIdentifier of relation '$identifier'")
             catalog.dropTable(tableIdentifier)
         }
     }
@@ -355,7 +362,7 @@ class HiveTableRelation(
 
         val missingFields = sourceSchema.filterNot(f => targetFields.contains(f.name.toLowerCase(Locale.ROOT)))
         if (missingFields.nonEmpty) {
-            logger.info(s"Migrating HiveTable relation '$identifier' by adding new columns ${missingFields.map(_.name).mkString(",")} to Hive table '$tableIdentifier'")
+            logger.info(s"Migrating HiveTable relation '$identifier' by adding new columns ${missingFields.map(_.name).mkString(",")} to Hive table $tableIdentifier")
             catalog.addTableColumns(tableIdentifier, missingFields)
         }
     }