+++ flowman-core/src/main/scala/com/dimajix/flowman/spec/model/HiveUnionTableRelation.scala
@@ -36,7 +36,7 @@ import com.dimajix.flowman.transforms.UnionTransformer
 import com.dimajix.flowman.types.FieldValue
 import com.dimajix.flowman.types.SingleValue
 import com.dimajix.flowman.util.SchemaUtils
-import com.dimajix.spark.sql.catalyst.SQLBuilder
+import com.dimajix.spark.sql.catalyst.SqlBuilder
 
 
 class HiveUnionTableRelation(
@@ -116,7 +116,7 @@ class HiveUnionTableRelation(
         val union = UnionTransformer().transformDataFrames(df)
         val finalSchema = schema.get.sparkSchema.fields ++ partitions.map(_.sparkField)
         val conformed = SchemaEnforcer(StructType(finalSchema)).transform(union)
-        val sql = new SQLBuilder(conformed).toSQL
+        val sql = new SqlBuilder(conformed).toSQL
         viewRelationFromSql(sql)
     }
 
@@ -171,7 +171,7 @@ class HiveUnionTableRelation(
         require(schema != null)
         require(partitions != null)
 
-        logger.info(s"Reading from Hive union view '$viewIdentifier' using partition values $partitions")
+        logger.info(s"Reading from Hive UNION VIEW $viewIdentifier using partition values $partitions")
 
         val tableDf = executor.spark.read.table(viewIdentifier.unquotedString)
         val df = filterPartition(tableDf, partitions)
@@ -200,7 +200,10 @@ class HiveUnionTableRelation(
                 val table = catalog.getTable(id)
                 SchemaUtils.isCompatible(df.schema, table.schema)
             }
-            .getOrElse(throw new ExecutionException(s"Cannot find appropriate target table for Hive Union Table '$identifier'"))
+            .getOrElse {
+                logger.error(s"Cannot find appropriate target table for Hive Union Table '$identifier'. Required schema is\n${df.schema.treeString}")
+                throw new ExecutionException(s"Cannot find appropriate target table for Hive Union Table '$identifier'")
+            }
 
         // 3. Drop  partition from all other tables
         allTables.filter(_ != table).foreach { table =>
@@ -258,7 +261,7 @@ class HiveUnionTableRelation(
             // Create initial view
             val spark = executor.spark
             val df = spark.read.table(hiveTableRelation.tableIdentifier.unquotedString)
-            val sql = new SQLBuilder(df).toSQL
+            val sql = new SqlBuilder(df).toSQL
             val hiveViewRelation = viewRelationFromSql(sql)
             hiveViewRelation.create(executor, false)
         }
@@ -277,14 +280,13 @@ class HiveUnionTableRelation(
             val catalog = executor.catalog
 
             // Destroy view
-            val view = viewIdentifier
-            logger.info(s"Dropping union view '$view' from Hive Union Table '$identifier'")
-            catalog.dropView(view, false)
+            logger.info(s"Dropping UNION VIEW $viewIdentifier from Hive Union Table relation '$identifier'")
+            catalog.dropView(viewIdentifier, false)
 
             // Destroy tables
             listTables(executor)
                 .foreach { table =>
-                    logger.info(s"Dropping backend table '$table' from Hive Union Table '$identifier'")
+                    logger.info(s"Dropping backend Hive table '$table' from Hive Union Table relation '$identifier'")
                     catalog.dropTable(table, false, true)
                 }
         }
@@ -325,7 +327,7 @@ class HiveUnionTableRelation(
 
                 val missingFields = sourceSchema.filterNot(f => targetFields.contains(f.name.toLowerCase(Locale.ROOT)))
                 if (missingFields.nonEmpty) {
-                    logger.info(s"Migrating HiveUntionTable relation '$identifier' by adding new columns ${missingFields.map(_.name).mkString(",")} to Hive table '$id'")
+                    logger.info(s"Migrating Hive Untion Table relation '$identifier' by adding new columns ${missingFields.map(_.name).mkString(",")} to Hive table $id")
                     catalog.addTableColumns(id, missingFields)
                     true
                 }
@@ -338,7 +340,7 @@ class HiveUnionTableRelation(
                 //  3.2 Create new table
                 val tableSet = allTables.toSet
                 val version = (1 to 100000).find(n => !tableSet.contains(tableIdentifier(n))).get
-                logger.info(s"Migrating HiveUntionTable relation '$identifier' by creating new Hive table '${tableIdentifier(version)}'")
+                logger.info(s"Migrating Hive Untion Table relation '$identifier' by creating new Hive table ${tableIdentifier(version)}")
                 val hiveTableRelation = tableRelation(version)
                 hiveTableRelation.create(executor, false)
                 true